Fix UI issues when switching band profiles

+++ app.rb
@@ -435,19 +435,33 @@ end
 
 get '/songs/:id' do
   require_login
-  @song = current_band.songs.find(params[:id])
+  @song = current_band.songs.find_by(id: params[:id])
+  
+  unless @song
+    redirect '/songs'
+  end
+  
   erb :show_song
 end
 
 get '/songs/:id/edit' do
   require_login
-  @song = current_band.songs.find(params[:id])
+  @song = current_band.songs.find_by(id: params[:id])
+  
+  unless @song
+    redirect '/songs'
+  end
+  
   erb :edit_song
 end
 
 put '/songs/:id' do
   require_login
-  @song = current_band.songs.find(params[:id])
+  @song = current_band.songs.find_by(id: params[:id])
+  
+  unless @song
+    redirect '/songs'
+  end
   
   if @song.update(params[:song])
     redirect "/songs/#{@song.id}"
@@ -459,15 +473,18 @@ end
 
 delete '/songs/:id' do
   require_login
-  song = current_band.songs.find(params[:id])
+  song = current_band.songs.find_by(id: params[:id])
   
-  # Clean up associations before deleting the song
-  song.set_list_songs.destroy_all
+  if song
+    # Clean up associations before deleting the song
+    song.set_list_songs.destroy_all
+    
+    # Remove the song from all bands (many-to-many relationship)
+    song.band_ids = []
+    
+    song.destroy
+  end
   
-  # Remove the song from all bands (many-to-many relationship)
-  song.band_ids = []
-  
-  song.destroy
   redirect '/songs'
 end
 
@@ -631,21 +648,35 @@ end
 
 get '/set_lists/:id' do
   require_login
-  @set_list = filter_by_current_band(SetList).includes(:venue).find(params[:id])
+  @set_list = filter_by_current_band(SetList).includes(:venue).find_by(id: params[:id])
+  
+  unless @set_list
+    redirect '/set_lists'
+  end
+  
   @available_songs = filter_by_current_band(Song).where.not(id: @set_list.song_ids).order(:title)
   erb :show_set_list
 end
 
 get '/set_lists/:id/edit' do
   require_login
-  @set_list = filter_by_current_band(SetList).find(params[:id])
+  @set_list = filter_by_current_band(SetList).find_by(id: params[:id])
+  
+  unless @set_list
+    redirect '/set_lists'
+  end
+  
   @venues = filter_by_current_band(Venue).order(:name)
   erb :edit_set_list
 end
 
 put '/set_lists/:id' do
   require_login
-  @set_list = filter_by_current_band(SetList).find(params[:id])
+  @set_list = filter_by_current_band(SetList).find_by(id: params[:id])
+  
+  unless @set_list
+    redirect '/set_lists'
+  end
   set_list_params = {
     name: params[:name], 
     notes: params[:notes],
@@ -667,16 +698,28 @@ end
 
 delete '/set_lists/:id' do
   require_login
-  set_list = filter_by_current_band(SetList).find(params[:id])
-  set_list.destroy
+  set_list = filter_by_current_band(SetList).find_by(id: params[:id])
+  
+  if set_list
+    set_list.destroy
+  end
+  
   redirect '/set_lists'
 end
 
 # Add song to set list
 post '/set_lists/:id/songs' do
   require_login
-  set_list = filter_by_current_band(SetList).find(params[:id])
-  song = filter_by_current_band(Song).find(params[:song_id])
+  set_list = filter_by_current_band(SetList).find_by(id: params[:id])
+  
+  unless set_list
+    redirect '/set_lists'
+  end
+  song = filter_by_current_band(Song).find_by(id: params[:song_id])
+  
+  unless song
+    redirect '/set_lists'
+  end
   position = set_list.set_list_songs.count + 1
   
   set_list_song = SetListSong.new(
@@ -697,7 +740,11 @@ end
 # Remove song from set list
 delete '/set_lists/:set_list_id/songs/:song_id' do
   require_login
-  set_list = filter_by_current_band(SetList).find(params[:set_list_id])
+  set_list = filter_by_current_band(SetList).find_by(id: params[:set_list_id])
+  
+  unless set_list
+    redirect '/set_lists'
+  end
   set_list_song = set_list.set_list_songs.find_by(song_id: params[:song_id])
   set_list_song.destroy if set_list_song
   
@@ -712,14 +759,23 @@ end
 # Print set list
 get '/set_lists/:id/print' do
   require_login
-  @set_list = filter_by_current_band(SetList).find(params[:id])
+  @set_list = filter_by_current_band(SetList).find_by(id: params[:id])
+  
+  unless @set_list
+    redirect '/set_lists'
+  end
   erb :print_set_list, layout: false
 end
 
 # Reorder songs in set list
 post '/set_lists/:id/reorder' do
   require_login
-  set_list = filter_by_current_band(SetList).find(params[:id])
+  set_list = filter_by_current_band(SetList).find_by(id: params[:id])
+  
+  unless set_list
+    content_type :json
+    return { success: false, error: "Set list not found" }.to_json
+  end
   song_order = params[:song_order]
   
   if song_order && song_order.is_a?(Array)
@@ -737,7 +793,11 @@ end
 post '/set_lists/:id/copy' do
   require_login
   begin
-    original_set_list = filter_by_current_band(SetList).find(params[:id])
+    original_set_list = filter_by_current_band(SetList).find_by(id: params[:id])
+    
+    unless original_set_list
+      redirect '/set_lists'
+    end
     
     # Create new set list with copied name and notes
     new_name = "Copy - #{original_set_list.name}"
@@ -969,7 +1029,12 @@ get '/venues/:id' do
   require_login
   return redirect '/set_lists' unless current_band
   
-  @venue = filter_by_current_band(Venue).find(params[:id])
+  @venue = filter_by_current_band(Venue).find_by(id: params[:id])
+  
+  unless @venue
+    redirect '/venues'
+  end
+  
   @bands = user_bands
   erb :show_venue
 end
@@ -978,7 +1043,12 @@ get '/venues/:id/edit' do
   require_login
   return redirect '/set_lists' unless current_band
   
-  @venue = filter_by_current_band(Venue).find(params[:id])
+  @venue = filter_by_current_band(Venue).find_by(id: params[:id])
+  
+  unless @venue
+    redirect '/venues'
+  end
+  
   erb :edit_venue
 end
 
@@ -986,7 +1056,11 @@ put '/venues/:id' do
   require_login
   return redirect '/set_lists' unless current_band
   
-  @venue = filter_by_current_band(Venue).find(params[:id])
+  @venue = filter_by_current_band(Venue).find_by(id: params[:id])
+  
+  unless @venue
+    redirect '/venues'
+  end
   if @venue.update(params[:venue])
     redirect "/venues/#{@venue.id}"
   else
@@ -999,8 +1073,12 @@ delete '/venues/:id' do
   require_login
   return redirect '/set_lists' unless current_band
   
-  venue = filter_by_current_band(Venue).find(params[:id])
-  venue.destroy
+  venue = filter_by_current_band(Venue).find_by(id: params[:id])
+  
+  if venue
+    venue.destroy
+  end
+  
   redirect '/venues'
 end
 
@@ -1056,7 +1134,11 @@ get '/venues/:venue_id/copy' do
   require_login
   return redirect '/set_lists' unless current_band
   
-  @venue = filter_by_current_band(Venue).find(params[:venue_id])
+  @venue = filter_by_current_band(Venue).find_by(id: params[:venue_id])
+  
+  unless @venue
+    redirect '/venues'
+  end
   
   # Get other bands the user is a member of
   @target_bands = current_user.bands.where.not(id: current_band.id).order(:name)
@@ -1073,7 +1155,11 @@ post '/venues/:venue_id/copy' do
   require_login
   return redirect '/set_lists' unless current_band
   
-  @venue = filter_by_current_band(Venue).find(params[:venue_id])
+  @venue = filter_by_current_band(Venue).find_by(id: params[:venue_id])
+  
+  unless @venue
+    redirect '/venues'
+  end
   target_band_id = params[:target_band_id]
   
   if target_band_id.blank?
